requirements.txt
dj-database-url==0.5.0
Django==3.2.13
psycopg2-binary==2.9.3
python-decouple==3.6
django-crispy-forms==1.14
django-debug-toolbar==3.4
django-filter==21.1
django-bootstrap4==22.1
django-widget-tweaks==1.4.12
django-extensions==3.1.5
jupyter==1.0.0
python-dateutil==2.8.2
django-q==1.3.9
django-localflavor==3.1
django-login-required-middleware==0.8.0
django-adminlte3==0.1.6
sentry-sdk==1.12.1
django-q[sentry]
pandas==1.5.2
openpyxl==3.0.10
pyexcel-xlsx==0.6.0